move filedata to new types module for narrower importability
